# storage/api-docs/api-docs-v1.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Notebook Api",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/notebook": {
            "get": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Get List Notebook Entries",
                "operationId": "98206cd0bb2b8eb30cfa0bc87555aab7",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "full_name": {
                                                        "type": "string",
                                                        "example": "Sergei Rachmaninoff"
                                                    },
                                                    "company": {
                                                        "type": "string",
                                                        "example": "Apple"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "+7987654321"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "format": "email",
                                                        "example": "email@email.com"
                                                    },
                                                    "date_of_birth": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "1985-03-19"
                                                    },
                                                    "photo_path": {
                                                        "type": "string",
                                                        "example": "photo.png"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Create Notebook Entry",
                "operationId": "5debba6b80d7436a76e0a688f9659d82",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "full_name": {
                                                "type": "string"
                                            },
                                            "company": {
                                                "type": "string"
                                            },
                                            "phone": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string",
                                                "format": "email"
                                            },
                                            "date_of_birth": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "photo_path": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "full_name": "Sergei Rachmaninoff",
                                    "company": "Apple",
                                    "phone": "+7987654321",
                                    "email": "email@email.com",
                                    "date_of_birth": "1985-03-19",
                                    "photo_path": "photo.png"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "full_name": {
                                            "type": "string",
                                            "example": "Sergei Rachmaninoff"
                                        },
                                        "company": {
                                            "type": "string",
                                            "example": "Apple"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "+7987654321"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "email@email.com"
                                        },
                                        "date_of_birth": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "1985-03-19"
                                        },
                                        "photo_path": {
                                            "type": "string",
                                            "example": "photo.png"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/notebook/{id}": {
            "get": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Get Detail Notebook Entry",
                "operationId": "31d2700b6f9c140103705f16c2eabb0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "full_name": {
                                            "type": "string",
                                            "example": "Sergei Rachmaninoff"
                                        },
                                        "company": {
                                            "type": "string",
                                            "example": "Apple"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "+7987654321"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "email@email.com"
                                        },
                                        "date_of_birth": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "1985-03-19"
                                        },
                                        "photo_path": {
                                            "type": "string",
                                            "example": "photo.png"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Update Notebook Entry",
                "operationId": "ad86d7b3344cd714e3042f1d14d1efa7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "full_name": {
                                                "type": "string"
                                            },
                                            "company": {
                                                "type": "string"
                                            },
                                            "phone": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "date_of_birth": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "photo_path": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "full_name": "Sergei Rachmaninoff",
                                    "company": "Apple",
                                    "phone": "+7987654321",
                                    "email": "email@email.com",
                                    "date_of_birth": "1985-03-19",
                                    "photo_path": "photo.png"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "full_name": {
                                            "type": "string",
                                            "example": "Sergei Rachmaninoff"
                                        },
                                        "company": {
                                            "type": "string",
                                            "example": "Apple"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "+7987654321"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "email@email.com"
                                        },
                                        "date_of_birth": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "1985-03-19"
                                        },
                                        "photo_path": {
                                            "type": "string",
                                            "example": "photo.png"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Delete Notebook Entry",
                "operationId": "d5fedcf194a2d56010d663dafaa69319",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "delete notebook entry success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}